backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
reportlab
pillow
psycopg2
requests